requirements.txt
pandas>=1.5.0
numpy>=1.24.0
yfinance>=0.2.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
